base config setup

+++ src/main/java/net/thedudemc/ninjabot/event/CommandEvents.java
@@ -6,6 +6,7 @@ import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.thedudemc.ninjabot.command.base.BotCommand;
 import net.thedudemc.ninjabot.command.exception.InvalidCommandException;
 import net.thedudemc.ninjabot.init.BotCommands;
+import net.thedudemc.ninjabot.init.BotConfigs;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.TimeUnit;
@@ -17,6 +18,8 @@ public class CommandEvents extends ListenerAdapter {
         String content = event.getMessage().getContentRaw();
         if (!content.startsWith("-")) return; // not a command.
 
+        final boolean deleteCommands = BotConfigs.getConfig("General").getBoolean("deleteCommands");
+
         try {
             String commandName = content.contains(" ") ? content.substring(1, content.indexOf(" ")) : content.substring(1);
             BotCommand command = BotCommands.getCommand(commandName);
@@ -27,9 +30,16 @@ public class CommandEvents extends ListenerAdapter {
                 command.execute(event.getMember(), event.getChannel(), event.getMessage(), args);
             }
         } catch (InvalidCommandException exception) {
-            event.getChannel().sendMessage(new MessageBuilder(exception.getMessage()).build()).queue(message -> message.delete().queueAfter(10, TimeUnit.SECONDS));
+            event.getChannel().sendMessage(new MessageBuilder(exception.getMessage()).build())
+                    .queue(message -> {
+                        if (deleteCommands) {
+                            message.delete().queueAfter(10, TimeUnit.SECONDS);
+                        }
+                    });
         } finally {
-            event.getMessage().delete().queueAfter(2, TimeUnit.SECONDS);
+            if (deleteCommands) {
+                event.getMessage().delete().queueAfter(2, TimeUnit.SECONDS);
+            }
         }
 
     }
